requirements.txt
python-telegram-bot==22.6
aiosqlite>=0.19
python-dotenv>=1.0
pytz>=2024.1
